requirements.txt
# LangGraph Multi-Agent System Requirements

# Core LangGraph dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# LLM providers
google-generativeai>=0.8.0
transformers>=4.45.0
torch>=2.0.0
accelerate>=0.34.0
bitsandbytes>=0.43.0

# Vector databases and embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0
duckdb>=1.1.0

# Data processing
pandas>=2.0.0
pyarrow>=17.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

# Security
cryptography>=41.0.0
keyring>=24.0.0

# Logging and monitoring
structlog>=23.0.0

# Optional: Web interface
streamlit>=1.28.0